services/mishka-llm-provider/pyproject.toml
[project]
name = "mishka-llm-provider"
version = "0.1.0"
description = "LLM Provider service for Mishka AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "google-generativeai",
    "python-dotenv",
    "httpx"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "httpx"
]
